refactor(lotto): rearrange the positions of the classes
