+++ src/main/java/lotto/controller/Validator.java
@@ -1,4 +1,7 @@
-package lotto;
+package lotto.controller;
+
+import lotto.utils.Constant;
+import lotto.utils.ExceptionMessage;
 
 public class Validator {
     public static void isNumber(String number){
