+++ src/main/java/lotto/view/InputView.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
+import lotto.utils.Constant;
 
 public class InputView {
     public static String getTicketCount(){
